Added support for multiple files
Adds support for the file chooser to select multiple files and to combine them into the tree node heirarchy.

OpenapiTreenodeConverterService has taken the parsing logic out from the api path tree component and now holds state which is updated for each file selected. It has a "reset" method which is used when a new set of files are loaded to clear previous state.

https://github.com/dhutchison/OpenApiVisualiser/issues/1